+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
   - `archetype` to create a Lume archetype in the local folder `_archetype/[name].ts`.
   - `cms` to create the `_cms.ts` file.
 - `feed` plugin: added `authorAvatar` property [#859].
+- RAM usage info in the debugbar.
 
 ### Changed
 - The `merge` utility function changed the types to better reflect its returned type.